requirements.txt

---

## 🧾 requirements.txt

pandas
numpy
scikit-learn
matplotlib
joblib
scipy
apache-airflow
openpyxl
